A few clean-ups, tests passing
