+++ tests/test_sources.py
@@ -5,10 +5,10 @@
 import fdtd
 import pytest
 
-from fixtures import grid, pml, periodic_boundary, backend_parametrizer
-from sources import SoftArbitraryPointSource
+from fixtures import grid, pml, periodic_boundary
+from fdtd.sources import SoftArbitraryPointSource
 ## Tests
 
 
 def test_SoftArbitraryPointSource(grid):
-    j
+    pass
